fix: correctly fall back to English name in LanguageSelector
